Don't instantiate serializers for simple types
Use static methods. I originally didn't want to do this because I though
it would be consistent to always have class instances act as
serializers, but then I decided that the difference between a serializer
with a parameter (one that needs to be instantiated) and a serializer
without a parameter is explicit and should be explicitly visible.

This change should also allow me to use class names as models and
abolish the S object from cosmic.